Add files via upload
The game starts with a square board consisting of tiles numbered from 1 to N and one empty tile, represented by the number 0. The goal is to arrange the tiles in numerical order. Moves are made by sliding tiles into the empty space from the top, bottom, left, or right.

The input consists of:

An integer N representing the number of numbered tiles (8, 15, 24, etc.).
An integer I, which is the index of the zero tile in the goal state (if -1 is provided, the default position is set to the bottom-right corner).
The initial board configuration.
Using the IDA* algorithm and the Manhattan distance heuristic, the program should output:

On the first line, the length of the "optimal" path from the initial state to the goal state.
The sequence of steps (one per line) required to reach the goal state. The steps should be: left, right, up, or down.
Notes:
Not every given puzzle configuration is solvable. A puzzle's solvability can be checked using the method described [here]. If the puzzle is unsolvable, the output should be -1 (i.e., no solution).
There should be an option to print the time taken to find the solution (excluding the time to print the output), measured in seconds with at least two decimal places. 

Example Input:
8  
-1  
1 2 3  
4 5 6  
0 7 8  

Example Output:
2  
left  
left